Add configuration option `IgnoreMetadata`
