+++ spec/support/expect_offense.rb
@@ -24,4 +24,17 @@ module ExpectOffense
   def inspected_source_filename
     DEFAULT_FILENAME
   end
+
+  def expect_global_offense(source, file = nil, message = '')
+    processed_source = parse_source(source, file)
+    offenses = _investigate(cop, processed_source)
+    expect(offenses.size).to eq(1)
+    expect(offenses.first.message).to eq(message)
+  end
+
+  def expect_no_global_offenses(source, file = nil)
+    processed_source = parse_source(source, file)
+    offenses = _investigate(cop, processed_source)
+    expect(offenses.size).to eq(0)
+  end
 end
